Deep-Live-Cam/remote-swap-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# Pin for RunPod PyTorch CUDA 11.8 images; 1.19+ often targets CUDA 12 and falls back to CPU.
onnxruntime-gpu==1.17.1
onnx>=1.16.0
numpy>=1.23.5,<2
opencv-python-headless>=4.8.0
python-multipart>=0.0.9
